Modified all controller to show error page for accessing actions with non-admin account, which is only accessible from admin accounts.

+++ CompanyCard/Controllers/CompaniesController.cs
@@ -27,7 +27,7 @@ namespace CompanyCard.Controllers
             }
             else
             {
-                return PartialView("ErrorLogin", new ErrorViewModel { Description = "Your must login first!!." });
+                return View("ErrorLogin", new ErrorViewModel { Description = "Your must login first!!." });
             }
 
         }
@@ -37,11 +37,18 @@ namespace CompanyCard.Controllers
         {
             if (Session["username"] != null)
             {
-                return PartialView();
+                if (Session["Admin"].Equals("Yes"))
+                {
+                    return PartialView();
+                }
+                else
+                {
+                    return PartialView("Error", new ErrorViewModel { Description = "Only Admins can do this!!." });
+                }
             }
             else
             {
-                return PartialView("ErrorLogin", new ErrorViewModel { Description = "Your must login first!!." });
+                return View("ErrorLogin", new ErrorViewModel { Description = "Your must login first!!." });
             }
         }
 
@@ -54,18 +61,25 @@ namespace CompanyCard.Controllers
         {
             if (Session["username"] != null)
             {
-                if (ModelState.IsValid)
+                if (Session["Admin"].Equals("Yes"))
                 {
-                    db.Companies.Add(company);
-                    db.SaveChanges();
-                    return RedirectToAction("Index");
-                }
+                    if (ModelState.IsValid)
+                    {
+                        db.Companies.Add(company);
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
 
-                return PartialView("Create", company);
+                    return PartialView("Create", company);
+                }
+                else
+                {
+                    return PartialView("Error", new ErrorViewModel { Description = "Only Admins can do this!!." });
+                }
             }
             else
             {
-                return PartialView("ErrorLogin", new ErrorViewModel { Description = "Your must login first!!." });
+                return View("ErrorLogin", new ErrorViewModel { Description = "Your must login first!!." });
             }
         }
 
@@ -74,23 +88,30 @@ namespace CompanyCard.Controllers
         {
             if (Session["username"] != null)
             {
-                if (id == null)
+                if (Session["Admin"].Equals("Yes"))
                 {
-                    return PartialView("Error", new ErrorViewModel { Description = "You should select a company." });
+                    if (id == null)
+                    {
+                        return PartialView("Error", new ErrorViewModel { Description = "You should select a company." });
+                    }
+                    Company company = db.Companies.Find(id);
+                    if (company == null)
+                    {
+                        return PartialView("Error", new ErrorViewModel { Description = "Company not found." });
+                    }
+                    //if (Request.IsAjaxRequest()) {
+                    //    return PartialView("Edit",company);
+                    //}
+                    return PartialView("Edit", company);
                 }
-                Company company = db.Companies.Find(id);
-                if (company == null)
+                else
                 {
-                    return PartialView("Error", new ErrorViewModel { Description = "Company not found." });
+                    return PartialView("Error", new ErrorViewModel { Description = "Only Admins can do this!!." });
                 }
-                //if (Request.IsAjaxRequest()) {
-                //    return PartialView("Edit",company);
-                //}
-                return PartialView("Edit", company);
             }
             else
             {
-                return PartialView("ErrorLogin", new ErrorViewModel { Description = "Your must login first!!." });
+                return View("ErrorLogin", new ErrorViewModel { Description = "Your must login first!!." });
             }
         }
 
@@ -103,17 +124,24 @@ namespace CompanyCard.Controllers
         {
             if (Session["username"] != null)
             {
-                if (ModelState.IsValid)
+                if (Session["Admin"].Equals("Yes"))
                 {
-                    db.Entry(company).State = EntityState.Modified;
-                    db.SaveChanges();
-                    return RedirectToAction("Index");
+                    if (ModelState.IsValid)
+                    {
+                        db.Entry(company).State = EntityState.Modified;
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
+                    return PartialView("Edit", company);
                 }
-                return PartialView("Edit", company);
+                else
+                {
+                    return PartialView("Error", new ErrorViewModel { Description = "Only Admins can do this!!." });
+                }
             }
             else
             {
-                return PartialView("ErrorLogin", new ErrorViewModel { Description = "Your must login first!!." });
+                return View("ErrorLogin", new ErrorViewModel { Description = "Your must login first!!." });
             }
         }
 
@@ -122,16 +150,23 @@ namespace CompanyCard.Controllers
         {
             if (Session["username"] != null)
             {
-                if (id == null)
+                if (Session["Admin"].Equals("Yes"))
                 {
-                    return PartialView("Error", new ErrorViewModel { Description = "You should select a company." });
+                    if (id == null)
+                    {
+                        return PartialView("Error", new ErrorViewModel { Description = "You should select a company." });
+                    }
+                    Company company = db.Companies.Find(id);
+                    if (company == null)
+                    {
+                        return View("Error", new ErrorViewModel { Description = "Company not found." });
+                    }
+                    return PartialView("Delete", company);
                 }
-                Company company = db.Companies.Find(id);
-                if (company == null)
+                else
                 {
-                    return View("Error", new ErrorViewModel { Description = "Company not found." });
+                    return PartialView("Error", new ErrorViewModel { Description = "Only Admins can do this!!." });
                 }
-                return PartialView("Delete", company);
             }
             else
             {
@@ -146,10 +181,17 @@ namespace CompanyCard.Controllers
         {
             if (Session["username"] != null)
             {
-                Company company = db.Companies.Find(id);
-                db.Companies.Remove(company);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (Session["Admin"].Equals("Yes"))
+                {
+                    Company company = db.Companies.Find(id);
+                    db.Companies.Remove(company);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    return PartialView("Error", new ErrorViewModel { Description = "Only Admins can do this!!." });
+                }
             }
             else
             {
